LayoutTests/fast/events/ios/meaningful-click-when-tapping-draggable-content.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<style>
body, html {
    font-size: 18px;
    font-family: system-ui;
    width: 100%;
    height: 100%;
    margin: 0;
}

#target {
    width: 100%;
    height: 100%;
    background-color: silver;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test exercises the 'meaningful click' heuristic when tapping draggable content.");

    if (!window.testRunner)
        return;

    const target = document.getElementById("target");
    target.addEventListener("mousedown", () => { });

    async function simulateTap(expectMeaningfulClick) {
        didHandleTap = false;
        didHandleTapAsMeaningfulClick = false;
        testRunner.installDidHandleTapCallback(wasMeaningful => {
            didHandleTap = true;
            didHandleTapAsMeaningfulClick = wasMeaningful;
        });

        await UIHelper.activateElement(target);
        await shouldBecomeEqual("didHandleTap", "true");
        if (expectMeaningfulClick)
            shouldBeTrue("didHandleTapAsMeaningfulClick");
        else
            shouldBeFalse("didHandleTapAsMeaningfulClick");
        testRunner.clearTestRunnerCallbacks();
        await UIHelper.waitForDoubleTapDelay();
    }

    await simulateTap(true);

    target.removeAttribute("draggable");
    await simulateTap(false);

    finishJSTest();
});
</script>
</head>
<body>
    <div id="description"></div>
    <div id="console"></div>
    <div id="target" draggable="true"></div>
</body>
</html>
