LayoutTests/fast/events/ios/meaningful-click-when-tapping-draggable-content-expected.txt

This test exercises the 'meaningful click' heuristic when tapping draggable content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS didHandleTap became true
PASS didHandleTapAsMeaningfulClick is true
PASS didHandleTap became true
PASS didHandleTapAsMeaningfulClick is false
PASS successfullyParsed is true

TEST COMPLETE
